Removing explicit initialization patch to iri_sub.
Rather than loading the IG and Ap/F10.7 tables on each invocation of
iri_sub, this functionality has been exposed to the user for the sake of
computationally efficiency (initialization is required only once and not
on each invocation). The trade-off is added complexity to the user, for
now the code will yield a garbage result if read_ig_rz and readapf107
are not first called before invoking iri_sub.