Use cache of geophysical_indices to avoid spin-up time.
Update geophysical_indices only when a index file has been updated /
added (as evident by the file modification time).

This adds state context to the code (i.e., added complexity) but the
execution time is greatly reduced (though this is amortized on multiple
invocations of Point). For my use cases, this is a net win and a
worthwhile optimization.

Race conditions exist with this implementation --- the cache could get
out of whack if an update occurred during the step in which file
modification times are queried. This code make the implicit assumption
that the user would not do such a thing.