Python3 compatibility (#50)
* Updated codes to be python3 compatible.

The end of python2 is coming. Updated the code with the help of
futurize to make compatible with python3.

* Fix for library file names --- darwin only!

Need to add a hack for other platforms. Current version will bomb on
other platforms.

* Added required absolute import for python 3.

* Must open in binary mode on python3.

* Fixed lingering binary issue snafus.

Python 2 and 3 handle binary / text strings differently.

* Change print statements to function calls.

* Updated codes to be python3 compatible.

The end of python2 is coming. Updated the code with the help of
futurize to make compatible with python3.

* Fix for library file names --- darwin only!

Need to add a hack for other platforms. Current version will bomb on
other platforms.

* Added required absolute import for python 3.

* Must open in binary mode on python3.

* Fixed lingering binary issue snafus.

Python 2 and 3 handle binary / text strings differently.

* Change print statements to function calls.

* Python 3 for more platforms (#47)

* build extensions in setup.py

* edit Makefiles to allow setup without compiling

* remove leftover comment

* correct encoding for hwm07e.f90

* make source before compile in each individual Makefile

* Remember to check for typos before pushing...

* Reinstate perl command for hwm07

* Updated README instructions

* Get rid of ipdb package