Update to "official" version of HWM14
Douglas Drob brought to our attention that pyglow had been using a beta version of HWM14. However, the beta version is almost identical to the offical version: winds are identical to within 10^-9 m/s. This commit updates to the official version of the software that is attached to Drob et al (2015).